修改部分代码，增加登录、创建用户等功能

+++ client/input_handler_register.go
@@ -28,7 +28,7 @@ func (c *Client) Transport(id messageId.MessageId, message proto.Message) {
 		return
 	}
 	c.cli.ChMsg <- &network.Message{
-		ID:   uint64(id),
+		ID:   id,
 		Data: bytes,
 	}
 }
